- Replaced the "Time Today" metric with "Positions reviewed today" (uses existing `DailyActivity.positionsReviewed`). No DB migrations required. - Backend: `GET /api/training-stats` now returns `positionsReviewedToday` (aggregated for UTC "today"). - Frontend: `HomePanel` shows "reviews today" and listens for optimistic `training-stats-updated` events; `TrainingClient` dispatches `{ positionsReviewed: 1 }` on successful reviews. - Repertoire `LineTree` now previews only the last two moves (no leading ellipsis) to better fit mobile. - Tests updated: API, HomePanel, and LineTree tests adjusted; local test run: 12 suites, 132 tests passed. - Added CI deploy workflow: `.github/workflows/deploy.yml`.
